Extension code refactoring
